Update README.md
This week I worked mainly on polishing the visualizations and helping them get implemented into the site.  This was finalizng the zoom functionality, color scaling and overall visual of all three prop viz's. Additionally oliver and I spent a lot of time getting the program to run two versions of D3 until vercel required the use of only one so I converted the map to use D3 v4 instead of v7.