week 4 git update - dan
This week I worked on merging everyone's updates into a cohesive whole (that ideally deploys on vercel 🤞) and ensuring the export functionality I had been building in the latter half of week 3 would work with all this new componentry. We managed to get everything working while locally hosted but ran into a lot of type safety errors when trying to deploy to our production site with Vercel. We'll have it up and running by block 7!